Styled Origin card at optimizePage.tsx
