+++ src/pages/OptimizePage/OptimizePage.tsx
@@ -14,10 +14,11 @@ import MapSection from '../../components/MapSection/MapSection';
 import PlacesForVisit from '../../components/PlacesForVisit/PlacesForVisit';
 import theme from '../../theme/theme';
 import SuggestedResultAccordion from '../../components/Accordion/SuggestedResultAccordion';
-import {StyledViewMapButton, StyledRemoveButton} from './Style';
+import {StyledViewMapButton, StyledOriginCard} from './Style';
 import JumpButton from '../../components/JumpButton/JumpButton';
 import { optimizedPlacesAction } from '../../features/optimizedPlaces/optimizedPlaceSlice';
 import SaveItineraryPopButton from '../../components/SaveItineraryPopup/SaveItineraryPopButton';
+import AddMoreActivitiesCard from '../../components/AddMoreActivities/AddMoreActivitiesCard'
 
 
 interface IActivity {
@@ -264,28 +265,25 @@ const OptimizePage = () => {
   }
   
   return (
-    <div style={{marginTop: "4rem"}}>
-      <Box sx={{height: "2rem", width: "100%", backgroundColor: theme.palette.primary.variant}}>
+    <div>
+      {/* <Box sx={{height: "2rem", width: "100%", backgroundColor: theme.palette.primary.variant}}>
         <img src="images/Helper-Text-2.png" alt="helper-text" style={{marginLeft: "10%"}}/>
-      </Box>
-      <Box sx={{display: "flex", justifyContent: "center", alignItems: "center"}}>
-        <SuggestedResultAccordion />
-      </Box>
+      </Box> */}
 
-      <Box sx={{width: "100%", display: "flex", justifyContent: "center", alignItems: "center", mt: "0.5rem"}}>
+      {/* <Box sx={{width: "100%", display: "flex", justifyContent: "center", alignItems: "center", mt: "0.5rem"}}>
         <img src="/images/Progress-3.png" alt="loading-bar" />
-      </Box>
+      </Box> */}
 
       <DragDropContext onDragEnd={handleDragAndDrop}>
-        <Grid container>
-          <Grid item laptop={6} sx={{paddingLeft: "5%"}} >
+        {/* <Box> */}
+          {/* <Box> */}
 
-            <Paper  elevation={3} sx={{width: "95%", marginBottom: "2rem", marginTop: "2rem", p: "1rem"}}>
+            <StyledOriginCard>
               <OriginCard {...origin} />
 
               <SaveItineraryPopButton />
               
-            </Paper>
+            </StyledOriginCard>
 
             <div style={{marginBottom: "1rem"}}>
               { dailyGroups && 
@@ -342,80 +340,17 @@ const OptimizePage = () => {
                 </div>
               )}
             </Droppable>
-          </Grid>
+          {/* </Box> */}
 
-          <Grid item laptop={6}>
-            
-            <Box sx={{mt: "2rem"}}>
-              <Typography variant="h6" component="div" sx={{mb: "1rem"}}>
-                Add More Activities
-              </Typography>
-              <Stack spacing={2} direction="row" sx={{border: "2px black solid", borderRadius: 99, padding: "0.5rem", mb:"1rem"}} width={"90%"}>
-                  <div>
-                    <SearchIcon sx={{color: "gray", marginTop: "0.3rem"}}/>
-                  </div>
-                  <div style={{width: "100%"}}>
-                    <Autocomplete onLoad={onLoad} onPlaceChanged={onPlaceChanged}>
-                      <InputBase type="search" placeholder="Search restaurants, attractions and more" sx={{width: "95%"}}/>
-                    </Autocomplete>
-                  </div>
-                                        
-              </Stack>
-
-              {newActivity && 
-                <Droppable droppableId="ADD-ACTIVITY">
-                {(provided: DroppableProvided) => (
-                  <div ref={provided.innerRef} {...provided.droppableProps} style={{height: "15vh", marginBottom: "1rem"}}>
-
-                    {newActivity.items.length > 0 && newActivity.items.map((item: any, index: number) => {
-                    
-                    return (
-                      <div> 
-                      <Draggable draggableId={item.location_id} index={index}>
-                        {(provided) => (
-                          <div ref={provided.innerRef} {...provided.draggableProps} {...provided.dragHandleProps}>
-                            <Grid container spacing={2} sx={{display: "flex", justifyContent: "baseline", alignItems: "center"}}>
-                              <Grid item laptop={10} >
-                                <Card sx={{width: "95%", mt: "0.5rem", px: "0.5rem" }} >                        
-                                  <CardContent>
-                                    <Box sx={{display: "flex", justifyContent: "space-between", flexDirection: "column"}}>
-                                      <Typography  variant="body1" component="div">
-                                        {item.name}
-                                      </Typography>
-
-                                      {item.address && <Typography variant="caption">
-                                        <span style={{fontWeight: "bold"}}>Address:</span> {" "} {item.address}
-                                      </Typography>}
-                                    </Box>
-                                  </CardContent> 
-                                </Card>
-                              </Grid>
-                              <Grid item laptop={2}>
-                                <Box>
-                                  <StyledRemoveButton 
-                                    aria-label="remove"
-                                    onClick={() => setNewActivity(null)}
-                                  >
-                                    <RemoveIcon fontSize="small"/>
-                                  </StyledRemoveButton>
-                                </Box>
-                              </Grid>
-                            </Grid>
- 
-                          </div>
-                        
-                        )}
-                      </Draggable>
-                      </div>
-                        
-                        ) })}
-                    {provided.placeholder}
-                  </div>
-                )}
-                </Droppable>
-          
-              }
-              
+          {/* <Box> */}
+            <Box>
+              <AddMoreActivitiesCard 
+              onLoad={onLoad}
+              onPlaceChanged={onPlaceChanged}
+              newActivity={newActivity}
+              setNewActivity={setNewActivity}
+              Autocomplete={Autocomplete}
+              />
             </Box>
             
             <Box sx={{width: "100%"}}>
@@ -433,8 +368,8 @@ const OptimizePage = () => {
               />
             </Box>
           
-          </Grid>
-        </Grid>
+          {/* </Box> */}
+        {/* </Box> */}
       </DragDropContext>
     </div>
   )
