+++ src/pages/OptimizePage/Style.ts
@@ -1,5 +1,5 @@
 import { styled } from '@mui/material/styles';
-import { Button } from '@mui/material';
+import { Button, Paper } from '@mui/material';
 import theme from '../../theme/theme';
 
 
@@ -28,16 +28,7 @@ export const StyledViewMapButton = styled(Button)({
   },
 });
 
-export const StyledRemoveButton = styled(Button)({
-  variant: 'contained',
-  backgroundColor: theme.palette.secondary.variant,
-  color: "#000000",
-  borderColor: '#000000',
-  borderRadius: 99,
-  '&:hover': {
-    backgroundColor: theme.palette.secondary.main,    
-    boxShadow: '1px 1px 1px 1px #000000',
-  },
-  textTransform: 'none',
-  boxShadow: "1px 1px 1px 1px #000000"
-});
+export const StyledOriginCard = styled(Paper)({
+  margin: "5%",
+  padding: "3%",
+})
